Import the initial basics of the forum
Currently this can show categories, topics, and posts. Post formatting
is rendered using the s9e/text-formatter package, with intermediate XML
stored in the database.

Pomm is being used to talk to a Postgresql database. I'm currently using
a forked copy of pomm-project/pomm-bundle with a patch applied so it
works with Twig 3. Once install/schema.sql is imported, there is a
console command (dev:generate-sample-data) to generate some sample data.